fix: create sub package note: this introduces a reformat of generate script closes: #6 changelog: - wip: major shift for an attempt to make generation scripts cleaner and better organised - package will use component so component can be isolate source to deal with src, view and test files
- fix: reorganise generate script for a cleaner logic
