+++ .config
@@ -1,0 +1,3 @@
+# this is a configuration file that has variables set to be used in various scripts 
+
+GLOBAL_PREFIX=pap
