+++ bin/generate/component/template/default/src/.config
@@ -1,0 +1,5 @@
+# Config file for "PLACEHOLDER_NAME"
+
+NAME=PLACEHOLDER_NAME
+CLASS_NAME=PLACEHOLDER_CLASS_NAME
+HTML_NAME=PLACEHOLDER_HTML_NAME
