+++ bin/generate/component/template/default/src/style.scss
