+++ bin/generate/component/template/default/src/types.ts
