+++ bin/generate/component/template/default/test/snapshot.test.ts
