+++ bin/generate/component/template/default/views/main.js
@@ -1,0 +1,9 @@
+// tools
+import '@papit/web';
+
+// component
+import 'PLACEHOLDER_PACKAGE_NAME';
+
+window.onload = () => {
+  console.log('[demo]: window loaded');
+}
