+++ bin/generate/component/template/default/views/public/translations/en.json
@@ -1,0 +1,7 @@
+{
+  "meta": {
+    "region": "GB",
+    "language": "en-UK"
+  },
+  "PLACEHOLDER_NAME": "PLACEHOLDER_NAME"
+}
