+++ bin/generate/component/template/default/views/style.css
