+++ bin/generate/package/template/.config
@@ -1,0 +1,16 @@
+# variables 
+ROOTDIR_RELATIVE=PLACEHOLDER_ROOTDIR_RELATIVE
+
+# layer info 
+LAYER_NAME=PLACEHOLDER_LAYER_NAME
+LAYER_FOLDER=PLACEHOLDER_LAYER_FOLDER
+
+# package info 
+PACKAGE_NAME=PLACEHOLDER_PACKAGE_NAME
+FULL_NAME=PLACEHOLDER_FULL_NAME
+
+# main component 
+NAME=PLACEHOLDER_NAME
+CLASS_NAME=PLACEHOLDER_CLASS_NAME
+
+# components
