+++ bin/generate/package/template/.gitignore
