+++ bin/generate/package/template/bin/_utils/init.sh
