+++ bin/generate/package/template/bin/_utils/react.sh
