+++ bin/generate/package/template/bin/build.sh
