+++ bin/generate/package/template/bin/executable.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+# NOTE: by deafult calling `npx PLACEHOLDER_PACKAGE_NAME`
+# will execute this file, as defined in package.json under the "bin" section
+
+echo "  *  <-- this is an asshole"
