+++ bin/generate/package/template/bin/watch.sh
