+++ bin/generate/package/template/tsconfig.json
