+++ bin/generate/package/template/tsconfig.prod.json
