+++ bin/generate/project/run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+echo "project generate run"
