--- bin/generate/sub-package.sh
@@ -1,1 +1,0 @@
-#!/bin/bash
