--- bin/generate/template/package/.env
@@ -1,12 +1,0 @@
-# variables 
-ROOTDIR_RELATIVE=TEMPLATE_ROOTDIR_RELATIVE
-LAYER_TYPE=TEMPLATE_LAYER_TYPE
-NAME=TEMPLATE_NAME
-PACKAGE_NAME=TEMPLATE_PACKAGE_NAME
-CLASS_NAME=TEMPLATE_CLASS_NAME
-HTML_NAME=TEMPLATE_HTML_NAME
-FULL_NAME=TEMPLATE_FULL_NAME
-
-# component 
-TEMPLATE_CLASS_NAME=src/component.ts
-TEMPLATE_CLASS_NAME_folder=src/
