--- bin/generate/template/package/bin/executable.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-# NOTE: by deafult calling `npx TEMPLATE_FULL_NAME`
-# will execute this file, as defined in package.json under the "bin" section
-
-echo "  *  <-- this is an asshole"
