--- bin/generate/template/package/src/index.ts
@@ -1,15 +1,0 @@
-import { TEMPLATE_CLASS_NAME } from './component.js';
-
-// export 
-export * from "./component";
-export * from "./types";
-
-// Register the element with the browser
-
-if (!window.customElements) {
-  throw new Error('Custom Elements not supported');
-}
-
-if (!window.customElements.get('TEMPLATE_HTML_NAME')) {
-  window.customElements.define('TEMPLATE_HTML_NAME', TEMPLATE_CLASS_NAME);
-}
