--- bin/generate/template/package/views/prototype/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>TEMPLATE_CLASS_NAME prototype</title>
-
-  <link rel="icon" type="image/x-icon" href="favicon.ico">
-
-  <!-- Styles -->
-  <link rel="stylesheet" href="style.css" />
-
-  <script defer src="main.js"></script>
-</head>
-
-<body>
-  <pap-card>
-    <pap-typography slot="header" variant="C1">Title</pap-typography>
-    <pap-codeblock display="both">
-      <TEMPLATE_HTML_NAME></TEMPLATE_HTML_NAME>
-    </pap-codeblock>
-  </pap-card>
-</body>
-
-</html>
