--- bin/generate/template/package/views/prototype/main.js
@@ -1,9 +1,0 @@
-// tools
-import '@papit/web';
-
-// component
-import 'TEMPLATE_FULL_NAME';
-
-window.onload = () => {
-  console.log('[demo]: window loaded');
-}
