--- bin/generate/template/package/views/prototype/public/translations/en.json
@@ -1,7 +1,0 @@
-{
-  "meta": {
-    "region": "GB",
-    "language": "en-UK"
-  },
-  "TEMPLATE_NAME": "TEMPLATE_NAME"
-}
