--- config.pap
@@ -1,3 +1,0 @@
-# this is a configuration file that has variables set to be used in various scripts 
-
-PREFIX=pap
